Improve code organisation
