+++ src/middleware/aws/instrumentation/fluent_builder/dynamodb.rs
@@ -2,7 +2,7 @@
 /// Generated by GitHub Copilot with human guidance
 use std::collections::HashSet;
 
-use super::{AwsInstrumentBuilder, utils::*};
+use super::{AwsInstrumentBuilder, instrument_aws_operation, utils::AsAttribute};
 use crate::{middleware::aws::*, semconv};
 
 impl<'a> AwsInstrumentBuilder<'a> for GetItemFluentBuilder {
