+++ src/middleware/aws/instrumentation/fluent_builder/firehose.rs
@@ -1,6 +1,6 @@
 /// AWS Firehose fluent builder instrumentation implementations
 /// Generated by GitHub Copilot with human guidance
-use super::{AwsInstrumentBuilder, utils::*};
+use super::{AwsInstrumentBuilder, instrument_aws_operation, utils::AsAttribute};
 use crate::{middleware::aws::*, semconv};
 
 // Publish operations
