+++ src/middleware/aws/instrumentation/fluent_builder/mod.rs
@@ -73,3 +73,38 @@ impl<'a, T: AwsInstrumentBuilder<'a>> InstrumentedFluentBuilder<'a, T> {
         self
     }
 }
+
+/// Generates [`super::InstrumentedFluentBuilder`] implementation for AWS SDK operations.
+macro_rules! instrument_aws_operation {
+    ($sdk:ident::operation::$op:ident, $builder:ident, $output:ident, $error:ident) => {
+        use $sdk::operation::$op::builders::$builder;
+        impl
+            super::InstrumentedFluentBuilder<'_, $sdk::operation::$op::builders::$builder>
+        {
+            /// Executes the AWS operation with instrumentation.
+            ///
+            /// This method creates a span for the operation and executes it within
+            /// that span context, providing automatic distributed tracing.
+            pub async fn send(
+                self,
+            ) -> Result<
+                $sdk::operation::$op::$output,
+                $sdk::error::SdkError<$sdk::operation::$op::$error>,
+            > {
+                self.inner.send().instrument(self.span).await
+            }
+        }
+    };
+    ($sdk:ident::operation::$op:ident) => {
+        paste::paste! {
+            instrument_aws_operation!(
+                $sdk::operation::$op,
+                [<$op:camel FluentBuilder>],
+                [<$op:camel Output>],
+                [<$op:camel Error>]
+            );
+        }
+    };
+}
+
+pub(super) use instrument_aws_operation;
